tree: extract derivate ref from flush
